fix: update demo to work on ghpages
