+++ examples/custom-tooltips-demo/index.html
@@ -4,6 +4,10 @@
 <head>
   <meta charset="utf-8">
   <title>Blockly Custom Tooltip Example</title>
+  <script src="./node_modules/blockly/blockly_compressed.js"></script>
+  <script src="./node_modules/blockly/msg/en.js"></script>
+  <script src="./node_modules/@blockly/dev-tools/dist/index.js"></script>
+  <script src="./index.js"></script>
   <style>
     html, body {
       margin: 0;
@@ -13,7 +17,6 @@
 
 <body>
   <div id="root"></div>
-  <script src="../build/test_bundle.js"></script>
 
   <xml xmlns="https://developers.google.com/blockly/xml" id="toolbox" style="display: none">
     <block type="custom_tooltip_1"></block>
