+++ examples/custom-tooltips-demo/index.js
@@ -8,9 +8,34 @@
  * @fileoverview Test page for example plugin showing custom tooltip rendering.
  */
 
-import * as Blockly from 'blockly';
-import {createPlayground} from '@blockly/dev-tools';
-import {CustomTooltips} from '../src/index';
+
+/**
+ * Create and register the custom tooltip rendering function.
+ * This could be extracted into a plugin if desired.
+ */
+function initTooltips() {
+  // Create a custom rendering function. This function will be called whenever
+  // a tooltip is shown in Blockly. The first argument is the div to render
+  // the content into. The second argument is the element to show the tooltip
+  // for.
+  const customTooltip = function(div, element) {
+    const tip = Blockly.Tooltip.getTooltipOfObject(element);
+    const text = document.createElement('div');
+    text.textContent = tip;
+    const container = document.createElement('div');
+    container.style.display = 'flex';
+    if (element.tooltipImg) {
+      const img = document.createElement('img');
+      img.setAttribute('src', element.tooltipImg);
+      container.appendChild(img);
+    }
+    container.appendChild(text);
+    div.appendChild(container);
+  };
+  // Register the function in the Blockly.Tooltip so that Blockly calls it
+  // when needed.
+  Blockly.Tooltip.setCustomTooltip(customTooltip);
+}
 
 /**
  * Create a workspace.
@@ -21,9 +46,7 @@ import {CustomTooltips} from '../src/index';
 function createWorkspace(blocklyDiv, options) {
   const workspace = Blockly.inject(blocklyDiv, options);
 
-  // Initialize the plugin.
-  const plugin = new CustomTooltips(workspace);
-  plugin.init();
+  initTooltips();
 
   return workspace;
 }
@@ -41,7 +64,7 @@ document.addEventListener('DOMContentLoaded', function() {
   Blockly.Blocks['custom_tooltip_2'] = {
     init: function() {
       this.appendDummyInput()
-          .appendField('This is a different test block.');
+          .appendField('Mouse over me.');
       this.setColour(150);
       this.setTooltip('Tip: This tooltip has an image.');
       // We will check for this property in our custom rendering code.
@@ -52,6 +75,7 @@ document.addEventListener('DOMContentLoaded', function() {
   const defaultOptions = {
     toolbox: document.getElementById('toolbox'),
   };
+  // eslint-disable-next-line no-undef
   createPlayground(document.getElementById('root'), createWorkspace,
       defaultOptions);
 });
