+++ examples/custom-tooltips-demo/package.json
@@ -1,39 +1,24 @@
 {
   "name": "custom-tooltips-demo",
   "version": "0.0.0",
-  "scripts": {
-    "build": "blockly-scripts build",
-    "clean": "blockly-scripts clean",
-    "start": "blockly-scripts start"
-  },
-  "main": "./dist/index.js",
-  "module": "./src/index.js",
-  "license": "Apache-2.0",
-  "directories": {
-    "dist": "dist",
-    "src": "src"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "devDependencies": {
+  "private": true,
+  "dependencies": {
     "@blockly/dev-scripts": "^1.2.14",
     "@blockly/dev-tools": "^3.0.7",
-    "blockly": "^7.20211209.4"
+    "showdown": "^2.0.0",
+    "blockly": "^8"
   },
-  "peerDependencies": {
-    "blockly": "^7.20211209.4"
-  },
-  "engines": {
-    "node": ">=8.17.0"
-  },
-  "browserslist": [
-    "defaults",
-    "IE 11",
-    "IE_Mob 11"
-  ],
-  "dependencies": {
-    "showdown": "^2.0.0"
+  "blocklyDemoConfig": {
+    "title": "Custom Tooltips",
+    "description": "Example of using a custom tooltip renderer. Make sure you select the 'default' toolbox.",
+    "files": [
+      "index.html",
+      "index.js",
+      "lightbulb.png",
+      "./node_modules/blockly/blockly_compressed.js",
+      "./node_modules/blockly/blockly_compressed.js.map",
+      "./node_modules/blockly/msg/en.js",
+      "./node_modules/@blockly/dev-tools/dist/index.js"
+    ]
   }
 }
